requirements.txt
# Health Tracker — Python dependencies
# Install with: pip install -r requirements.txt
# Compatible with Python 3.10+ on Windows, macOS, and Linux

# --- Core (required for daily sync) ---
garminconnect>=0.2.38
gspread>=6.2.1
google-auth>=2.49.0
python-dotenv>=0.21.0
keyring>=24.3.1
requests>=2.31.0

# --- Analytics (required for analysis_*.py and weekly_report.py) ---
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
